Assignment3/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
selenium==4.15.0
requests==2.31.0
